+++ src/Model/PacketaShipmentExporter.php
@@ -1,0 +1,238 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ThreeBRS\SyliusPacketaPlugin\Model;
+
+use Sylius\Component\Core\Model\OrderItemInterface;
+use Sylius\Component\Core\Model\PaymentInterface;
+use Sylius\Component\Core\Model\PaymentMethodInterface;
+use Sylius\Component\Core\Model\ShipmentInterface;
+use Sylius\Component\Currency\Converter\CurrencyConverter;
+use ThreeBRS\SyliusShipmentExportPlugin\Model\ShipmentExporterInterface;
+
+class PacketaShipmentExporter implements ShipmentExporterInterface
+{
+    /**
+     * @param array<string> $shippingMethodsCodes
+     */
+    public function __construct(
+        private readonly CurrencyConverter $currencyConverter,
+        private readonly array $shippingMethodsCodes,
+    ) {
+    }
+
+    private function convert(
+        int $amount,
+        string $sourceCurrencyCode,
+        string $targetCurrencyCode,
+    ): int {
+        return $this->currencyConverter->convert($amount, $sourceCurrencyCode, $targetCurrencyCode);
+    }
+
+    /**
+     * @return array<string>
+     */
+    public function getShippingMethodsCodes(): array
+    {
+        return $this->shippingMethodsCodes;
+    }
+
+    /**
+     * @param array<mixed> $questionsArray
+     *
+     * @return array<mixed>
+     */
+    public function getRow(
+        ShipmentInterface $shipment,
+        array $questionsArray,
+    ): array {
+        assert($shipment instanceof PacketaShipmentInterface);
+
+        $order = $shipment->getOrder();
+        assert($order !== null);
+        $channel = $order->getChannel();
+        assert($channel !== null);
+        $address = $order->getShippingAddress();
+        assert($address !== null);
+        $customer = $order->getCustomer();
+        assert($customer !== null);
+
+        $shippingMethod = $shipment->getMethod();
+        assert($shippingMethod instanceof PacketaShippingMethodInterface);
+        $payment = $order->getPayments()->first();
+        assert($payment instanceof PaymentInterface);
+        $paymentMethod = $payment->getMethod();
+        assert($paymentMethod instanceof PaymentMethodInterface);
+        assert($paymentMethod->getGatewayConfig() !== null);
+
+        $isCashOnDelivery = $paymentMethod->getGatewayConfig()->getFactoryName() === 'offline';
+
+        $currencyCode = $order->getCurrencyCode();
+        assert($currencyCode !== null);
+
+        $packeta = $shipment->getPacketa();
+        $targetCurrencyCode = null;
+        $decimals = 0;
+        $totalAmount = '';
+        if ($packeta !== null && array_key_exists('currency', $packeta) && $packeta['currency'] !== null) {
+            $targetCurrencyCode = $packeta['currency'];
+            assert(is_string($targetCurrencyCode));
+            $totalAmount = $this->convert($order->getTotal(), $currencyCode, $targetCurrencyCode);
+        }
+
+        if ($targetCurrencyCode === null) {
+            if ($packeta !== null && array_key_exists('country', $packeta)) {
+                $countryCode = $packeta['country'];
+                assert(is_string($countryCode));
+            } else {
+                $countryCode = $address->getCountryCode();
+            }
+
+            if ($countryCode !== null) {
+                $countryCode = strtoupper($countryCode);
+                if ($countryCode === 'CZ') {
+                    $targetCurrencyCode = 'CZK';
+                    $totalAmount = $this->convert($order->getTotal(), $currencyCode, $targetCurrencyCode);
+                } elseif ($countryCode === 'SK') {
+                    $targetCurrencyCode = 'EUR';
+                    $totalAmount = $this->convert($order->getTotal(), $currencyCode, $targetCurrencyCode);
+                    $decimals = 2;
+                } elseif ($countryCode === 'PL') {
+                    $targetCurrencyCode = 'PLN';
+                    $totalAmount = $this->convert($order->getTotal(), $currencyCode, $targetCurrencyCode);
+                } elseif ($countryCode === 'HU') {
+                    $targetCurrencyCode = 'HUF';
+                    $totalAmount = $this->convert($order->getTotal(), $currencyCode, $targetCurrencyCode);
+                } elseif ($countryCode === 'RO') {
+                    $targetCurrencyCode = 'RON';
+                    $totalAmount = $this->convert($order->getTotal(), $currencyCode, $targetCurrencyCode);
+                } else {
+                    $totalAmount = null;
+                }
+            }
+        }
+
+        if ($totalAmount !== null) {
+            $totalAmount = number_format(
+                ((int) $totalAmount) / 100,
+                $decimals,
+                '.',
+                '',
+            );
+        }
+
+        $weight = 0;
+        foreach ($order->getItems() as $item) {
+            /** @var OrderItemInterface $item */
+            $variant = $item->getVariant();
+            if ($variant !== null) {
+                $weight += $variant->getWeight() * $item->getQuantity();
+            }
+        }
+
+        $packetaId = $packeta !== null && array_key_exists('id', $packeta)
+            ? $packeta['id']
+            : null;
+        $carrierId = $shippingMethod->getPacketaConfig() !== null
+            ? $shippingMethod->getPacketaConfig()->getCarrierId()
+            : null;
+        $senderLabel = $shippingMethod->getPacketaConfig() !== null
+            ? $shippingMethod->getPacketaConfig()->getSenderLabel()
+            : null;
+
+        return [
+            /* 1 - version 5 */
+            '',
+
+            /* 2 - Číslo obj.* */
+            $order->getNumber(),
+
+            /* 3 - Jméno* */
+            $address->getFirstName(),
+
+            /* 4 - Příjmení* */
+            $address->getLastName(),
+
+            /* 5 - Firma */
+            $address->getCompany(),
+
+            /* 6 - E-mail** */
+            $customer->getEmail(),
+
+            /* 7 - Mobil** */
+            $address->getPhoneNumber(),
+
+            /* 8 - Dobírka */
+            $isCashOnDelivery
+                ? $totalAmount
+                : '',
+
+            /* 9 - Měna */
+            $targetCurrencyCode,
+
+            /* 10 - Hodnota **/
+            $totalAmount,
+
+            /* 11 - Hmotnost */
+            $weight,
+
+            /* 12 - Cílová pobočka* */
+            $packetaId ?? $carrierId,
+
+            /* 13 - Doména e-shopu*** */
+            $senderLabel,
+
+            /* 14 - Obsah 18+ */
+            '',
+
+            /* 15 - Plánovaný výdej */
+            '',
+
+            /* 16 - Ulice */
+            $packetaId
+                ? ''
+                : $address->getStreet(),
+
+            /* 17 - Č. domu */
+            '',
+
+            /* 18 - Obec */
+            $packetaId
+                ? ''
+                : $address->getCity(),
+
+            /* 19 - PSČ */
+            $packetaId
+                ? ''
+                : $address->getPostcode(),
+
+            /* 20 - Unique ID of the carrier pickup point */
+            '',
+        ];
+    }
+
+    public function getDelimiter(): string
+    {
+        return ',';
+    }
+
+    /**
+     * @return array<mixed>|null
+     */
+    public function getQuestionsArray(): ?array
+    {
+        return null;
+    }
+
+    /**
+     * @return array<array<string>>|null
+     */
+    public function getHeaders(): ?array
+    {
+        return [
+            ['version 5'],
+            [''],
+        ];
+    }
+}