+++ src/ThreeBRSSyliusPacketaPlugin.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ThreeBRS\SyliusPacketaPlugin;
+
+use Sylius\Bundle\CoreBundle\Application\SyliusPluginTrait;
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class ThreeBRSSyliusPacketaPlugin extends Bundle
+{
+    use SyliusPluginTrait;
+}