+++ tests/Application/src/Entity/ShippingMethod.php
@@ -2,16 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Tests\ThreeBRS\SyliusZasilkovnaPlugin\Entity;
+namespace Tests\ThreeBRS\SyliusPacketaPlugin\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
 use Sylius\Component\Core\Model\ShippingMethod as BaseShippingMethod;
-use ThreeBRS\SyliusZasilkovnaPlugin\Model\ZasilkovnaShippingMethodInterface;
-use ThreeBRS\SyliusZasilkovnaPlugin\Model\ZasilkovnaShippingMethodTrait;
+use ThreeBRS\SyliusPacketaPlugin\Model\PacketaShippingMethodInterface;
+use ThreeBRS\SyliusPacketaPlugin\Model\PacketaShippingMethodTrait;
 
 #[ORM\MappedSuperclass]
 #[ORM\Table(name: 'sylius_shipping_method')]
-class ShippingMethod extends BaseShippingMethod implements ZasilkovnaShippingMethodInterface
+class ShippingMethod extends BaseShippingMethod implements PacketaShippingMethodInterface
 {
-    use ZasilkovnaShippingMethodTrait;
+    use PacketaShippingMethodTrait;
 }