+++ tests/Behat/Context/Ui/Admin/ManagingShippingMethodContext.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Tests\ThreeBRS\SyliusZasilkovnaPlugin\Behat\Context\Ui\Admin;
+namespace Tests\ThreeBRS\SyliusPacketaPlugin\Behat\Context\Ui\Admin;
 
 use Behat\Behat\Context\Context;
 use Sylius\Behat\Service\NotificationCheckerInterface;
-use Tests\ThreeBRS\SyliusZasilkovnaPlugin\Behat\Page\Admin\ShippingMethod\UpdatePageInterface;
+use Tests\ThreeBRS\SyliusPacketaPlugin\Behat\Page\Admin\ShippingMethod\UpdatePageInterface;
 use Webmozart\Assert\Assert;
 
 final readonly class ManagingShippingMethodContext implements Context
@@ -14,59 +14,60 @@ final readonly class ManagingShippingMethodContext implements Context
     public function __construct(private UpdatePageInterface $updatePage)
     {
     }
+
     /**
-     * @When /^I change Zásilkovna Sender label to "([^"]*)"$/
+     * @When /^I change Packeta Sender label to "([^"]*)"$/
      */
-    public function iChangeZasilkovnaSenderLabelTo($arg1): void
+    public function iChangePacketaSenderLabelTo($arg1): void
     {
         $this->updatePage->changeInput('senderLabel', $arg1);
     }
 
     /**
-     * @Then /^the Zásilkovna Sender label for this shipping method should be "([^"]*)"$/
+     * @Then /^the Packeta Sender label for this shipping method should be "([^"]*)"$/
      */
-    public function theZasilkovnaSenderLabelForThisShippingMethodShouldBe($arg1): void
+    public function thePacketaSenderLabelForThisShippingMethodShouldBe($arg1): void
     {
         Assert::eq($this->updatePage->getSingleResourceFromPage('senderLabel'), $arg1);
     }
 
     /**
-     * @When /^I change Zásilkovna Carrier pickup point to "([^"]*)"$/
+     * @When /^I change Packeta Carrier pickup point to "([^"]*)"$/
      */
-    public function iChangeZasilkovnaCarrierIdTo($arg1): void
+    public function iChangePacketaCarrierIdTo($arg1): void
     {
         $this->updatePage->changeInput('carrierId', $arg1);
     }
 
     /**
-     * @Then /^the Zásilkovna Carrier pickup point for this shipping method should be "([^"]*)"$/
+     * @Then /^the Packeta Carrier pickup point for this shipping method should be "([^"]*)"$/
      */
-    public function theZasilkovnaCarrierIdForThisShippingMethodShouldBe($arg1): void
+    public function thePacketaCarrierIdForThisShippingMethodShouldBe($arg1): void
     {
         Assert::eq($this->updatePage->getSingleResourceFromPage('carrierId'), $arg1);
     }
 
     /**
-     * @When /^I change Zásilkovna api key to "([^"]*)"$/
+     * @When /^I change Packeta api key to "([^"]*)"$/
      */
-    public function iChangeZasilkovnaApiKeyTo(string $apiKey): void
+    public function iChangePacketaApiKeyTo(string $apiKey): void
     {
         $this->updatePage->changeApiKey($apiKey);
     }
 
     /**
-     * @Then /^the Zásilkovna api key for this shipping method should be "([^"]*)"$/
+     * @Then /^the Packeta api key for this shipping method should be "([^"]*)"$/
      */
-    public function theZasilkovnaApiKeyForThisShippingMethodShouldBe(string $apiKey): void
+    public function thePacketaApiKeyForThisShippingMethodShouldBe(string $apiKey): void
     {
         Assert::eq($this->updatePage->getSingleResourceFromPage('apiKey'), $apiKey);
     }
 
     /**
-     * @Then it should be shipped to Zásilkovna branch
+     * @Then it should be shipped to Packeta branch
      */
-    public function itShouldBeShippedToZasilkovnaBranch(): void
+    public function itShouldBeShippedToPacketaBranch(): void
     {
-        Assert::true($this->updatePage->iSeeZasilkovnaBranchInsteadOfShippingAddress());
+        Assert::true($this->updatePage->iSeePacketaBranchInsteadOfShippingAddress());
     }
 }